requirements.txt
pandas
scikit-learn
apache-airflow
psycopg2
